Adapt updates code to use workers
