+++ spyder/utils/tests/test_updates.py
