+++ spyder/utils/workers/tests/test_manager.py
@@ -1,0 +1,1 @@
+
